Add result_class to Servactory configuration
The commit adds a new attribute, result_class, to the Servactory configuration DSL, setup and factory classes. It allows checking and raising an error if the passed in value isn't a subclass of Servactory::Result. Also, an example of result_class configuration has been added to the application_service in examples.