Refactor result methods and enhance tests
The methods for success and failure results in the TestKit module have been slightly refactored for clarity. Instead of creating a new context within the success_for/failure_for calls, the context is created beforehand in a separate step. This change makes the methods easier to understand and also facilitates debugging. Furthermore, additional checks have been added in various tests to ensure the returned result is not only a kind of Servactory::Result but also an instance of it.