Disable RuboCop's ABC Size metric in several methods
In the 'getter_with' methods of the 'outputs' and 'internals' classes, and the 'register_prepare_option' method in the 'registrar' class, the RuboCop ABC Size metric has been disabled. This was done to relax the linting rules to fit the requirements of the current codebase without modifying the complexity of the methods.