Refactor 'work' variable to 'actor'
In the existing code, a variable named 'work' was inconsistently describing an instance of an actor. To make the code more readable and coherent, 'work' was renamed to 'actor' in multiple areas including info method and inside valid_with_matcher test. This change emphasizes that the object in question is indeed an actor.