Refactor exception and test messages in specs
The code changes involve minor refactoring in the exception messages and expectation comparisons of several test specs. Indentation for error messages is improved for better readability. Also, instead of comparing the method return value to nil, the 'be_nil' matcher is used, which is more semantic and idiomatic in RSpec tests.