Add option name to dynamic options initialization and error messages
The dynamic options now include the option name on initialization and in the error messages. Made changes to the `WorkOption` initialization and equivalent_with definition in the `must.rb` file, as well as updating error messages in the `multiple_of.rb` and relevant localization files. This makes the error messages more informative about which option caused the error.