Reorder condition checks in multiple_of.rb
The conditions in the 'multiple_of' file have been reorganized for efficiency. The condition check for 'zero' value of an option is now placed after confirming that the option value belongs to numeric, float, rational, or big decimal. Hence, this prevents unnecessary 'zero' value checks on non-applicable option value types.