Add NoteMatcher and replace description with note
This commit introduces a new matcher in the test kit, NoteMatcher, for handling the testing of a 'note' attribute. References to the 'description' attribute in various classes and modules, including the input and attribute registration features, have been replaced with 'note'. This change is also reflected in some example code and tests.