Add 'note' attribute to outputs and test it
The 'note' attribute has been added to the outputs in the 'output.rb', 'matchers.rb', and 'dsl.rb' files. This attribute was also added to the test kit for RSpec and incorporated into the relevant tests within the 'example1_spec.rb' file. The examples given in 'example1.rb' have been updated to reflect these changes.